+++ CHANGELOG.md
@@ -195,3 +195,4 @@
 - 2023-12-03: Test find email, README, chat, Makefile, +2
 - 2023-12-03: Docs campaign, discover, LICENSE, mod, +4
 - 2023-12-04: Cache follow up, users, contacts, frontier, +5
+- 2023-12-04: Init daemon, campaigns, discovery, sender, +6